# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zupply IA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  critical_stock: 'critical_stock',
  pending_order: 'pending_order',
  prediction: 'prediction',
  info: 'info',
} as const;

export interface Alert {
  id: number;
  type: AlertType;
  message: string;
  createdAt: string;
}

export interface SupplierSummary {
  id: number;
  name: string;
  totalSpent: number;
  ordersCount: number;
}

export interface MonthlySpend {
  month: string;
  amount: number;
}

export interface DashboardSummary {
  totalOrdersThisMonth: number;
  totalSpentThisMonth: number;
  activeSuppliers: number;
  criticalStockCount: number;
  pendingOrdersCount: number;
  cashBalance: number;
  recentAlerts: Alert[];
  topSuppliers: SupplierSummary[];
  monthlySpend: MonthlySpend[];
}

export interface Review {
  id: number;
  supplierId: number;
  rating: number;
  comment: string;
  reviewerName: string;
  createdAt: string;
}

export interface Supplier {
  id: number;
  name: string;
  category: string;
  contactName: string;
  phone: string;
  email: string;
  address?: string;
  rating: number;
  reviewCount: number;
  isActive: boolean;
  createdAt: string;
  reviews?: Review[];
}

export interface SupplierInput {
  name: string;
  category: string;
  contactName: string;
  phone: string;
  email: string;
  address?: string;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  reviewerName: string;
}

export interface Product {
  id: number;
  supplierId: number;
  supplierName: string;
  name: string;
  category: string;
  unit: string;
  pricePerUnit: number;
  minOrderQty: number;
  isAvailable: boolean;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  supplierName: string;
  quantity: number;
  unit: string;
  unitPrice: number;
  subtotal: number;
}

export interface Order {
  id: number;
  status: OrderStatus;
  totalAmount: number;
  supplierCount: number;
  createdAt: string;
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  notes?: string | null;
  items?: OrderItem[];
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  notes?: string | null;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type InventoryItemStockStatus = typeof InventoryItemStockStatus[keyof typeof InventoryItemStockStatus];


export const InventoryItemStockStatus = {
  ok: 'ok',
  low: 'low',
  critical: 'critical',
} as const;

export interface InventoryItem {
  id: number;
  /** @nullable */
  productId?: number | null;
  name: string;
  category: string;
  unit: string;
  currentStock: number;
  minStock: number;
  maxStock: number;
  stockStatus: InventoryItemStockStatus;
  lastUpdated: string;
  avgDailyUsage: number;
}

export interface InventoryUpdate {
  currentStock?: number;
  minStock?: number;
  maxStock?: number;
}

export type InventoryPredictionUrgency = typeof InventoryPredictionUrgency[keyof typeof InventoryPredictionUrgency];


export const InventoryPredictionUrgency = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type InventoryPredictionTrend = typeof InventoryPredictionTrend[keyof typeof InventoryPredictionTrend];


export const InventoryPredictionTrend = {
  increasing: 'increasing',
  stable: 'stable',
  decreasing: 'decreasing',
} as const;

export interface InventoryPrediction {
  inventoryItemId: number;
  itemName: string;
  currentStock: number;
  unit: string;
  predictedDemand: number;
  recommendedOrderQty: number;
  /** @nullable */
  daysUntilStockout?: number | null;
  urgency: InventoryPredictionUrgency;
  recommendation: string;
  trend: InventoryPredictionTrend;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  expense: 'expense',
  income: 'income',
  adjustment: 'adjustment',
} as const;

export interface Transaction {
  id: number;
  /** @nullable */
  orderId?: number | null;
  type: TransactionType;
  amount: number;
  description: string;
  category: string;
  date: string;
}

export interface ExpenseCategory {
  category: string;
  amount: number;
  percentage: number;
}

export interface AccountingSummary {
  cashBalance: number;
  totalExpensesThisMonth: number;
  totalExpensesLastMonth: number;
  expensesByCategory: ExpenseCategory[];
  monthlyTrend: MonthlySpend[];
}

export type ListProductsParams = {
/**
 * @nullable
 */
supplierId?: number | null;
/**
 * @nullable
 */
category?: string | null;
};